+++ src/csv_tools/rss_ringoccs_get_geo.c
@@ -75,7 +75,7 @@ rssringoccs_Get_Geo(const char *filename, tmpl_Bool use_deprecated)
     int ch;
 
     unsigned int column_count = 0U;
-    unsigned long int n = 0UL;
+    size_t n = 0;
 
     /*  Allocate memory for the geo data.                                     */
     geo = malloc(sizeof(*geo));
