+++ src/fresnel_transform/rss_ringoccs_fresnel_transform_newton_norm.c
@@ -26,11 +26,10 @@
 void
 rssringoccs_Fresnel_Transform_Newton_Norm(rssringoccs_TAUObj *tau,
                                           double *w_func,
-                                          unsigned long n_pts,
-                                          unsigned long center)
+                                          size_t n_pts, size_t center)
 {
     /*  Declare all necessary variables. i and j are used for indexing.       */
-    unsigned long m, offset;
+    size_t m, offset;
 
     /*  The Fresnel kernel and the stationary ring azimuth angle.             */
     double psi, phi, cos_psi, sin_psi, real_norm, abs_norm;
