+++ src/fresnel_transform/rss_ringoccs_fresnel_transform_perturbed_newton_norm.c
@@ -26,11 +26,10 @@
 void
 rssringoccs_Fresnel_Transform_Perturbed_Newton_Norm(rssringoccs_TAUObj *tau,
                                                     double *w_func,
-                                                    unsigned long n_pts,
-                                                    unsigned long center)
+                                                    size_t n_pts, size_t center)
 {
     /*  Declare all necessary variables. i and j are used for indexing.       */
-    unsigned long m, offset;
+    size_t m, offset;
 
     /*  The Fresnel kernel and the stationary ring azimuth angle.             */
     double psi, phi, x, poly, cos_psi, sin_psi, abs_norm, real_norm;
